mcp_server/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
anthropic>=0.26.1
google-generativeai>=0.7.0
python-dotenv==1.0.1